Update README.md
This will ask the user how many toppings they want on their pizza, and then it will randomly generate a pizza with the specified number of toppings from the list of available options. The toppings will be displayed in a numbered list format using lists to store different toppings and the 'random' module to randomly select toppings for the pizza.